+++ include/pt/sampling/cosine_pdf.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "pt/math/onb.hpp"
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 #include "pt/sampling/pdf.hpp"
 
@@ -9,7 +10,7 @@ class cosine_pdf final : public pdf {
 public:
     explicit cosine_pdf(const vec3& normal) : uvw(normal) {}
 
-    [[nodiscard]] double value(const vec3& direction) const override;
+    [[nodiscard]] Float value(const vec3& direction) const override;
 
     [[nodiscard]] vec3 generate() const override;
