+++ include/pt/sampling/pdf.hpp
@@ -1,11 +1,12 @@
 #pragma once
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 
 namespace pt {
 
 class pdf {
 public:
-    [[nodiscard]] virtual double value(const vec3& direction) const = 0;
+    [[nodiscard]] virtual Float value(const vec3& direction) const = 0;
 
     [[nodiscard]] virtual vec3 generate() const = 0;
