+++ include/pt/sampling/sphere_pdf.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 #include "pt/sampling/pdf.hpp"
 
@@ -6,7 +7,7 @@ namespace pt {
 
 class sphere_pdf final : public pdf {
 public:
-    [[nodiscard]] double value(const vec3& direction) const override;
+    [[nodiscard]] Float value(const vec3& direction) const override;
 
     [[nodiscard]] vec3 generate() const override;
 };
