+++ include/pt/textures/noise_texture.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 #include "pt/textures/perlin.hpp"
 #include "pt/textures/texture.hpp"
@@ -7,13 +8,13 @@ namespace pt {
 
 class noise_texture final : public texture {
 public:
-    explicit noise_texture(double scale) : scale(scale) {}
+    explicit noise_texture(Float scale) : scale(scale) {}
 
-    [[nodiscard]] color value(double u, double v, const point3& p) const override;
+    [[nodiscard]] color value(Float u, Float v, const point3& p) const override;
 
 private:
     perlin noise;
-    double scale{};
+    Float scale{};
 };
 
 } // namespace pt
