+++ include/pt/textures/solid_color.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 #include "pt/textures/texture.hpp"
 
@@ -7,9 +8,9 @@ namespace pt {
 class solid_color final : public texture {
 public:
     explicit solid_color(const color& albedo) : albedo(albedo) {}
-    solid_color(double r, double g, double b) : solid_color(color(r, g, b)) {}
+    solid_color(Float r, Float g, Float b) : solid_color(color(r, g, b)) {}
 
-    [[nodiscard]] color value(double u, double v, const point3& p) const override;
+    [[nodiscard]] color value(Float u, Float v, const point3& p) const override;
 
 private:
     color albedo;
