+++ src/sampling/hittable_pdf.cpp
@@ -1,10 +1,11 @@
 #include "pt/sampling/hittable_pdf.hpp"
 #include "pt/core/hittable.hpp"
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 
 namespace pt {
 
-double hittable_pdf::value(const vec3& direction) const {
+Float hittable_pdf::value(const vec3& direction) const {
     return objects.pdf_value(origin, direction);
 }
