+++ src/textures/solid_color.cpp
@@ -1,8 +1,9 @@
 #include "pt/textures/solid_color.hpp"
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 
 namespace pt {
 
-color solid_color::value(double, double, const point3&) const { return albedo; }
+color solid_color::value(Float, Float, const point3&) const { return albedo; }
 
 } // namespace pt
